+++ evalml/pipelines/classification/catboost.py
@@ -10,7 +10,7 @@ class CatBoostClassificationPipeline(PipelineBase):
     Note: impute_strategy must support both string and numeric data
     """
     component_graph = ['Simple Imputer', 'CatBoost Classifier']
-    problem_types = ['binary', 'multiclass']
+    supported_problem_types = ['binary', 'multiclass']
     _hyperparameters = {
         "impute_strategy": ["most_frequent"],
     }