+++ evalml/pipelines/classification/random_forest.py
@@ -5,4 +5,4 @@ class RFClassificationPipeline(PipelineBase):
     """Random Forest Pipeline for both binary and multiclass classification"""
     _name = "Random Forest Classification Pipeline"
     component_graph = ['One Hot Encoder', 'Simple Imputer', 'RF Classifier Select From Model', 'Random Forest Classifier']
-    problem_types = ['binary', 'multiclass']
+    supported_problem_types = ['binary', 'multiclass']