+++ evalml/pipelines/regression/catboost.py
@@ -11,7 +11,7 @@ class CatBoostRegressionPipeline(PipelineBase):
     Note: impute_strategy must support both string and numeric data
     """
     component_graph = ['Simple Imputer', 'CatBoost Regressor']
-    problem_types = ['regression']
+    supported_problem_types = ['regression']
     _hyperparameters = {
         "impute_strategy": ["most_frequent"],
     }