+++ evalml/pipelines/regression/linear_regression.py
@@ -4,4 +4,4 @@ from evalml.pipelines import PipelineBase
 class LinearRegressionPipeline(PipelineBase):
     """Linear Regression Pipeline for regression problems"""
     component_graph = ['One Hot Encoder', 'Simple Imputer', 'Standard Scaler', 'Linear Regressor']
-    problem_types = ['regression']
+    supported_problem_types = ['regression']