+++ evalml/pipelines/regression/random_forest.py
@@ -5,4 +5,4 @@ class RFRegressionPipeline(PipelineBase):
     """Random Forest Pipeline for regression problems"""
     _name = "Random Forest Regression Pipeline"
     component_graph = ['One Hot Encoder', 'Simple Imputer', 'RF Regressor Select From Model', 'Random Forest Regressor']
-    problem_types = ['regression']
+    supported_problem_types = ['regression']